requirements.txt
# FastAPI and ASGI
fastapi==0.110.0
uvicorn==0.27.1

# MongoDB async driver
motor==3.3.2
pymongo==4.5.0  # ✅ Required for Motor

# Pydantic for data validation and email
pydantic[email]==2.6.3

# OpenAI API
openai==1.91.0  # ✅ Updated for compatibility with current client usage

# Environment variable support
python-dotenv==1.0.1

# CORS middleware
starlette==0.36.3

# File upload support
python-multipart==0.0.9  # ✅ Required for UploadFile & Form

# Optional (for local dev)
watchfiles==0.21.0

# If you serve static files with FastAPI
aiofiles==23.2.1
